Fix make_flux_sparsity_pattern on parallel meshes.
This fixes my erroneous work in 9b037871bb. The function previously
skipped faces based on the total ordering: however, we will not revisit
the face if it is between a locally owned and ghost cell, so the
resulting sparsity pattern is rather weird (and wrong). Instead, only
skip the face if we meet the previous ordering condition and both cells
are locally owned.