Get around a GCC 6 warning.
GCC 6.1 now warns (-Waddress) that (since the dynamic cast of an object
back to its own type always succeeds) the address of an object is always
nonzero, so these casts result in tautologies for DoFHandler instances
and PETSc MPI vectors.

Regardless of the warning, these assertions are useful for all
distributed vectors (and may be run now for sequential vectors too) and
dof handler type, so they should always be checked.